pre-parse reply to get choice blocks outgoingEdgeIds
