+++ packages/bot-engine/src/continueBotFlow.ts
@@ -5,12 +5,9 @@ import {
   isInputBlock,
 } from "@typebot.io/blocks-core/helpers";
 import type { Block } from "@typebot.io/blocks-core/schemas/schema";
-import { defaultChoiceInputOptions } from "@typebot.io/blocks-inputs/choice/constants";
 import { InputBlockType } from "@typebot.io/blocks-inputs/constants";
 import { defaultEmailInputOptions } from "@typebot.io/blocks-inputs/email/constants";
-import { defaultFileInputOptions } from "@typebot.io/blocks-inputs/file/constants";
 import { defaultPaymentInputOptions } from "@typebot.io/blocks-inputs/payment/constants";
-import { defaultPictureChoiceOptions } from "@typebot.io/blocks-inputs/pictureChoice/constants";
 import type { InputBlock } from "@typebot.io/blocks-inputs/schema";
 import { IntegrationBlockType } from "@typebot.io/blocks-integrations/constants";
 import { LogicBlockType } from "@typebot.io/blocks-logic/constants";
@@ -18,13 +15,10 @@ import type {
   SessionState,
   TypebotInSession,
 } from "@typebot.io/chat-session/schemas";
-import { env } from "@typebot.io/env";
 import { forgedBlocks } from "@typebot.io/forge-repository/definitions";
 import type { ForgedBlock } from "@typebot.io/forge-repository/schemas";
 import { getBlockById } from "@typebot.io/groups/helpers/getBlockById";
 import type { Group } from "@typebot.io/groups/schemas";
-import { parseAllowedFileTypesMetadata } from "@typebot.io/lib/extensionFromMimeType";
-import { isURL } from "@typebot.io/lib/isURL";
 import { parseUnknownError } from "@typebot.io/lib/parseUnknownError";
 import { byId, isDefined } from "@typebot.io/lib/utils";
 import type { Prisma } from "@typebot.io/prisma/types";
@@ -36,23 +30,16 @@ import type {
   SetVariableHistoryItem,
   Variable,
 } from "@typebot.io/variables/schemas";
-import { parseCardsReply } from "./blocks/cards/parseCardsReply";
-import { parseButtonsReply } from "./blocks/inputs/buttons/parseButtonsReply";
-import { parseDateReply } from "./blocks/inputs/date/parseDateReply";
-import { formatEmail } from "./blocks/inputs/email/formatEmail";
-import { parseNumber } from "./blocks/inputs/number/parseNumber";
-import { formatPhoneNumber } from "./blocks/inputs/phone/formatPhoneNumber";
-import { parsePictureChoicesReply } from "./blocks/inputs/pictureChoice/parsePictureChoicesReply";
-import { validateRatingReply } from "./blocks/inputs/rating/validateRatingReply";
-import { parseTime } from "./blocks/inputs/time/parseTime";
 import { saveDataInResponseVariableMapping } from "./blocks/integrations/httpRequest/saveDataInResponseVariableMapping";
 import { resumeChatCompletion } from "./blocks/integrations/legacy/openai/resumeChatCompletion";
 import { executeCommandEvent } from "./events/executeCommandEvent";
 import { executeReplyEvent } from "./events/executeReplyEvent";
 import { executeGroup, parseInput } from "./executeGroup";
 import { getNextGroup } from "./getNextGroup";
+import { getOutgoingEdgeId } from "./getOutgoingEdgeId";
 import { findReplyEvent } from "./helpers/findReplyEvent";
 import { isInputMessage } from "./helpers/isInputMessage";
+import { parseReply } from "./parseReply";
 import { saveAnswer } from "./queries/saveAnswer";
 import { resetSessionState } from "./resetSessionState";
 import { saveVariablesValueIfAny } from "./saveVariables";
@@ -78,6 +65,7 @@ type Params = {
   textBubbleContentFormat: "richText" | "markdown";
   sessionStore: SessionStore;
 };
+
 export const continueBotFlow = async (
   reply: Message | undefined,
   { state, version, startTime, textBubbleContentFormat, sessionStore }: Params,
@@ -553,184 +541,6 @@ const setNewAnswerInState =
       ),
     } satisfies SessionState;
   };
-
-const getOutgoingEdgeId = (
-  reply: SuccessReply | SkipReply | undefined,
-  {
-    block,
-    state,
-    sessionStore,
-  }: {
-    block: Block;
-    state: SessionState;
-    sessionStore: SessionStore;
-  },
-): { edgeId: string | undefined; isOffDefaultPath: boolean } => {
-  if (!reply || reply.status === "skip")
-    return { edgeId: block.outgoingEdgeId, isOffDefaultPath: false };
-  if (reply.outgoingEdgeId)
-    return { edgeId: reply.outgoingEdgeId, isOffDefaultPath: true };
-  const variables = state.typebotsQueue[0].typebot.variables;
-  if (
-    block.type === InputBlockType.CHOICE &&
-    !(
-      block.options?.isMultipleChoice ??
-      defaultChoiceInputOptions.isMultipleChoice
-    ) &&
-    reply
-  ) {
-    const matchedItem = block.items.find(
-      (item) =>
-        parseVariables(item.content, {
-          variables,
-          sessionStore,
-        }).normalize() === reply.content.normalize(),
-    );
-    if (matchedItem?.outgoingEdgeId)
-      return { edgeId: matchedItem.outgoingEdgeId, isOffDefaultPath: true };
-  }
-  if (
-    block.type === InputBlockType.PICTURE_CHOICE &&
-    !(
-      block.options?.isMultipleChoice ??
-      defaultPictureChoiceOptions.isMultipleChoice
-    ) &&
-    reply
-  ) {
-    const matchedItem = block.items.find(
-      (item) =>
-        parseVariables(item.title, { variables, sessionStore }).normalize() ===
-        reply.content.normalize(),
-    );
-    if (matchedItem?.outgoingEdgeId)
-      return { edgeId: matchedItem.outgoingEdgeId, isOffDefaultPath: true };
-  }
-  return { edgeId: block.outgoingEdgeId, isOffDefaultPath: false };
-};
-
-const parseReply = async (
-  reply: InputMessage | undefined,
-  {
-    sessionStore,
-    state,
-    block,
-  }: { sessionStore: SessionStore; state: SessionState; block: InputBlock },
-): Promise<ParsedReply> => {
-  switch (block.type) {
-    case InputBlockType.EMAIL: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      const formattedEmail = formatEmail(reply.text);
-      if (!formattedEmail) return { status: "fail" };
-      return { status: "success", content: formattedEmail };
-    }
-    case InputBlockType.PHONE: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      const formattedPhone = formatPhoneNumber(
-        reply.text,
-        block.options?.defaultCountryCode,
-      );
-      if (!formattedPhone) return { status: "fail" };
-      return { status: "success", content: formattedPhone };
-    }
-    case InputBlockType.URL: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      const isValid = isURL(reply.text, { require_protocol: false });
-      if (!isValid) return { status: "fail" };
-      return { status: "success", content: reply.text };
-    }
-    case InputBlockType.CHOICE: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parseButtonsReply(reply.text, {
-        block,
-        state,
-        sessionStore,
-      });
-    }
-    case InputBlockType.NUMBER: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parseNumber(reply.text, {
-        options: block.options,
-        variables: state.typebotsQueue[0].typebot.variables,
-        sessionStore,
-      });
-    }
-    case InputBlockType.DATE: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parseDateReply(reply.text, block);
-    }
-    case InputBlockType.TIME: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parseTime(reply.text, block.options);
-    }
-    case InputBlockType.FILE: {
-      if (!reply)
-        return (block.options?.isRequired ?? defaultFileInputOptions.isRequired)
-          ? { status: "fail" }
-          : { status: "skip" };
-
-      const replyValue = reply.type === "audio" ? reply.url : reply.text;
-      const urls = replyValue.split(", ");
-      const hasValidUrls = urls.some((url) =>
-        isURL(url, { require_tld: env.S3_ENDPOINT !== "localhost" }),
-      );
-
-      const allowedFileTypesMetadata =
-        block.options?.allowedFileTypes?.types &&
-        block.options?.allowedFileTypes?.types?.length > 0 &&
-        block.options?.allowedFileTypes?.isEnabled
-          ? parseAllowedFileTypesMetadata(block.options.allowedFileTypes.types)
-          : undefined;
-      const allFilesAreAllowed = allowedFileTypesMetadata
-        ? urls.every((url) => {
-            const extension = url.split(".").pop();
-            if (!extension) return false;
-            return allowedFileTypesMetadata.some(
-              (metadata) => metadata.extension === extension,
-            );
-          })
-        : true;
-
-      const status = hasValidUrls && allFilesAreAllowed ? "success" : "fail";
-      if (!block.options?.isMultipleAllowed && urls.length > 1)
-        return { status, content: replyValue.split(",")[0] };
-      return { status, content: replyValue };
-    }
-    case InputBlockType.PAYMENT: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      if (reply.text === "fail") return { status: "fail" };
-      return { status: "success", content: reply.text };
-    }
-    case InputBlockType.RATING: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      const isValid = validateRatingReply(reply.text, block);
-      if (!isValid) return { status: "fail" };
-      return { status: "success", content: reply.text };
-    }
-    case InputBlockType.PICTURE_CHOICE: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parsePictureChoicesReply(reply.text, {
-        block,
-        state,
-        sessionStore,
-      });
-    }
-    case InputBlockType.TEXT: {
-      if (!reply) return { status: "fail" };
-      return {
-        status: "success",
-        content: reply.type === "audio" ? reply.url : reply.text,
-      };
-    }
-    case InputBlockType.CARDS: {
-      if (!reply || reply.type !== "text") return { status: "fail" };
-      return parseCardsReply(reply.text, {
-        block,
-        state,
-        sessionStore,
-      });
-    }
-  }
-};
 
 export const safeJsonParse = (value: string): unknown => {
   try {
